pyproject.toml
[tool.poetry]
name = "django-drf-auth-profile"
version = "0.1.0"
description = ""
authors = ["valhallajazzy <valhallajazzy@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
python-dotenv = "^1.0.1"
environs = "^11.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
phonenumbers = "^8.13.35"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
